Switches root-5-dev requirements for rat-dev
Some issues with using root-5-dev on fresh install. Switching the order
here will download the latest release version of ROOT 5, for fresh
installs, and for when the release hasn't been installed.

If troubles with installation of ROOT 5 occur, then manually install
root-5-dev, and ensure root-5.34.36 is not installed. This will then
allow for for the development version to be used.